Improve GLS behaviour

+++ Assets/__Scripts/Utils/DataStructures/StateData.cs
@@ -9,7 +9,7 @@ public abstract class StateData<T> : IEquatable<StateData<T>> where T : BaseObje
     protected StateData(T data) => Base = data;
 
     public readonly T Base;
-    public float StartTime = float.MinValue;
+    public float StartTime = short.MinValue;
     public float EndTime = float.MaxValue;
 
     public bool Equals(StateData<T> other) => id == other!.id;
